requirements.txt
click~=8.1
pillow~=10.1
imageio~=2.30
opencv-python~=4.5
matplotlib~=3.8.3
scikit-image~=0.22.0
setuptools~=58.1.0
numpy~=1.26.4
scipy~=1.13.0